// Hackathon/2026/Steward/web/owner-console/src/api/types.ts
export type Network = 'test' | 'main'